pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "healthcare-ml-gcp-case-study"
version = "0.1.0"
description = "End-to-end healthcare ML case study using Google Cloud products"
readme = "README.md"
requires-python = ">=3.11"
authors = [{name = "Portfolio Owner"}]
dependencies = [
  "pandas>=2.2,<3",
  "numpy>=1.26,<3",
  "scikit-learn>=1.5,<2",
  "joblib>=1.4,<2",
  "pyyaml>=6.0,<7",
  "google-cloud-bigquery>=3.25,<4",
  "google-cloud-storage>=2.18,<3",
  "google-cloud-aiplatform>=1.71,<2",
  "kfp>=2.9,<3",
  "google-cloud-pipeline-components>=2.15,<3",
  "fastapi>=0.115,<1",
  "uvicorn>=0.30,<1"
]

[project.optional-dependencies]
dev = ["pytest>=8.3,<9"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
